Minor changes to flags/arguments (#2)
Use -e to check if file exist.
As far as I can tell -a is a binary operator to check that two expressions are true (source: man test)

Use x flag when opening nginx.conf for modification.
Since we move nginx.conf to nginx.conf.bak prior to running generate-conf.py - We never expect this to exist. Then it makes more sense to fail (x) if the file for some reason exist, instead of appending (a).