Add certbot-related targets to Makefile (#3)
Add three new certbot-related targets to Makefile. These has helped me to automate the process of obtaining and renew certificates for any https-server defined in servers.json.

validate-certs
Iterates the servers in servers.json, returns a non-zero exit status if any of the servers (with https: true) does not have fullchain.pem and privkey.pem files in the expected locations.

generate-certs
Iterates the servers in servers.json, attempting to generate a certificate with certbot for any server with https: true.

renew-certs
Attempts to renew any certificates previously obtained with certbot.